Move hashlib-compatible scrypt function tests
to a new test_hashlib_scrypt.py testfile
